requirements.txt
absl-py==0.9.0
astor==0.8.1
Click==7.0
cycler==0.10.0
Flask==1.1.1
gast==0.3.2
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.10.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.18.0
Pillow==7.0.0
pkg-resources==0.0.0
protobuf==3.11.2
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.13.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
